refactor: background.ts into multiple files to separate concern and improve readability and maintainability
