+++ src/background/handlers/handle-update-base-url.ts
@@ -1,0 +1,44 @@
+import type { SendResponseFunction } from "@/types"
+
+export async function handleUpdateBaseUrl(
+  payload: string,
+  sendResponse: SendResponseFunction
+): Promise<void> {
+  try {
+    const origin = new URL(payload).origin
+
+    await chrome.declarativeNetRequest.updateDynamicRules({
+      removeRuleIds: [1],
+      addRules: [
+        {
+          id: 1,
+          priority: 1,
+          action: {
+            type: chrome.declarativeNetRequest.RuleActionType.MODIFY_HEADERS,
+            requestHeaders: [
+              {
+                header: "Origin",
+                operation: chrome.declarativeNetRequest.HeaderOperation.SET,
+                value: origin
+              }
+            ]
+          },
+          condition: {
+            urlFilter: `${origin}/*`,
+            resourceTypes: [
+              chrome.declarativeNetRequest.ResourceType.XMLHTTPREQUEST
+            ]
+          }
+        }
+      ]
+    })
+
+    sendResponse({ success: true })
+  } catch (err) {
+    const error = err as Error
+    sendResponse({
+      success: false,
+      error: { status: 0, message: error.message }
+    })
+  }
+}
